Refactoring dep analysis.
- Move all dependency analysis code to the DependencyCache.
- Rearrange code into smaller reasonable units, to make
  it easier to understand.
